Factor out GeneralTab
